Synch the 2 pipelines lab schema
