+++ hillman_mouse_pipeline/+hillman_common_lab/Lab.m
@@ -1,0 +1,9 @@
+%{
+  lab                     : varchar(32)    # e.g. Randylab, Hillmanlab
+  ---
+  lab_description=''      : varchar(255)
+%}
+
+classdef Lab < dj.Lookup
+
+end
