+++ hillman_mouse_pipeline/+hillman_common_lab/Project.m
@@ -1,0 +1,10 @@
+%{
+  project                     : varchar(32)
+  ---
+  -> LabMember
+  project_description=''      : varchar(1024)
+%}
+
+classdef Project < dj.Lookup
+
+end
