+++ hillman_pipeline/experiment.py
@@ -5,37 +5,6 @@ schema = dj.schema('hillman_experiment')
 
 
 @schema
-class Lab(dj.Lookup):
-    definition = """
-    lab                     : varchar(32)    # e.g. Randylab, Hillmanlab
-    ---
-    lab_description=''      : varchar(255)
-    """
-    contents = [['hillmanlab', '']]
-
-
-@schema
-class LabMember(dj.Lookup):
-    definition = """
-    user        : varchar(32)
-    ---
-    -> Lab
-    """
-    contents = [['richardyan', 'hillmanlab'],
-                ['wenzeli', 'hillmanlab']]
-
-
-@schema
-class Project(dj.Lookup):
-    definition = """
-    project                     : varchar(32)
-    ---
-    -> LabMember
-    project_description=''      : varchar(1024)
-    """
-
-
-@schema
 class Species(dj.Lookup):
     definition = """
     species        :  varchar(32)
